requirements.txt
streamlit
pinecone-client
llama-index
llama_index-llms-gemini
llama-index-embeddings-gemini
llama-index-vector-stores-pinecone
BeautifulSoup4
google-generativeai